refactor: cargo hack fixes
