+++ crates/router/src/configs/settings.rs
@@ -495,7 +495,6 @@ pub struct Connectors {
     pub aci: ConnectorParams,
     #[cfg(feature = "payouts")]
     pub adyen: ConnectorParamsWithSecondaryBaseUrl,
-    #[cfg(feature = "payouts")]
     pub adyenplatform: ConnectorParams,
     #[cfg(not(feature = "payouts"))]
     pub adyen: ConnectorParams,
