+++ crates/router/tests/connectors/sample_auth.toml
@@ -230,4 +230,6 @@ api_key="API Key"
 
 [gpayments]
 api_key="API Key"
-\n\n[adyenplatform]\napi_key="API Key"
+
+[adyenplatform]
+api_key="API Key"
