+++ openapi/openapi_spec.json
@@ -16232,7 +16232,8 @@
           "return_url",
           "business_country",
           "description",
-          "entity_type"
+          "entity_type",
+          "priority"
         ],
         "properties": {
           "payout_id": {
@@ -16391,12 +16392,7 @@
             "nullable": true
           },
           "priority": {
-            "allOf": [
-              {
-                "$ref": "#/components/schemas/api_enums.PayoutSendPriority"
-              }
-            ],
-            "nullable": true
+            "$ref": "#/components/schemas/PayoutSendPriority"
           }
         },
         "additionalProperties": false
